[FE-1188] Addressing feedback
