+++ src/hooks/index.js
@@ -10,3 +10,4 @@ export { default as useProvidedContext } from './useProvidedContext';
 export { default as useSparkPostQuery } from './useSparkPostQuery';
 export { default as useTabs } from './useTabs';
 export { default as useUniqueId } from './useUniqueId';
+export { default as usePinnedReport } from './usePinnedRport';
